plugins/ui/alp-libs/python/pyodidepyqe/requirements.txt
requests==2.32.3
msal==1.31.0
pandas==1.5.3
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
